refactor: types에 있는 user타입 api폴더로 이동
