+++ src/app/auths/signup/page.tsx
@@ -3,7 +3,7 @@
 import Link from 'next/link';
 import React from 'react';
 import { SubmitHandler, useForm } from 'react-hook-form';
-import { SignUpFormData } from '@/types/user';
+import { SignUpFormData } from '@/api/auth/type';
 import useCreateUser from '@/hooks/api/auth/useCreateUser';
 import { useRouter } from 'next/navigation';
