--- src/types/api.ts
@@ -1,10 +1,0 @@
-export type ApiResponse<T> = {
-  data: T;
-  status: number;
-  message: string;
-};
-
-export type ApiError = {
-  status: number;
-  message: string;
-};
